add ability to pay for the order

+++ order-system/src/main/kotlin/Dish/Dish.kt
@@ -3,7 +3,7 @@ package Dish
 import kotlinx.serialization.Serializable
 
 @Serializable
-class Dish(var name: String, var number: Int, var price: Float, var complexity: Int) {
+class Dish(var name: String, var number: Int, var price: Double, var complexity: Int) {
     override fun toString(): String {
         return "Название: $name, кол-во: $number, цена: $price, сложность: $complexity"
     }
